+++ src/helpers/svgo-config.ts
@@ -4,14 +4,16 @@ import parsers from "../parsers";
 
 interface Params {
   readonly rawConfig: string;
-  readonly svgoVersion: number;
+  readonly svgoVersion: {
+    readonly value: number;
+  };
 }
 
 function parseRawSvgoConfig({
   rawConfig,
   svgoVersion,
 }: Params): [unknown, error] {
-  if (svgoVersion === 1) {
+  if (svgoVersion.value === 1) {
     const parseYaml = parsers.NewYaml();
     return parseYaml(rawConfig);
   } else {