+++ src/inputs/__mocks__/getters.ts
@@ -1,17 +1,19 @@
+import { defaultConfig } from "./__common__";
+
 const getIgnoreGlobs = jest.fn()
-  .mockReturnValue(["", null])
+  .mockReturnValue([defaultConfig.ignoreGlobs, null])
   .mockName("getters.getIgnoreGlobs");
 
 const getIsDryRun = jest.fn()
-  .mockReturnValue([false, null])
+  .mockReturnValue([defaultConfig.isDryRun, null])
   .mockName("getters.getIsDryRun");
 
 const getSvgoConfigPath = jest.fn()
-  .mockReturnValue(["", null])
+  .mockReturnValue([defaultConfig.svgoConfigPath, null])
   .mockName("getters.getSvgoConfigPath");
 
 const getSvgoVersion = jest.fn()
-  .mockReturnValue([2, null])
+  .mockReturnValue([defaultConfig.svgoVersion, null])
   .mockName("getters.getSvgoVersion");
 
 export {